Rename matrices_and_indices module and update imports
